requirements.txt
django==4.2.9
djangorestframework>=3.13.1
django-stronghold>=0.4.0
django-debug-toolbar>=3.6.0
zaproxy>=0.0.18
arabic-reshaper>=2.1.3
asgiref>=3.5.2
asn1crypto>=1.5.1
astroid>=2.11.7
async-generator>=1.10
attrs>=22.1.0
beautifulsoup4>=4.11.1
boto3>=1.24.56
botocore>=1.27.56
certifi>=2022.6.15
cffi>=1.15.1
chardet>=5.0.0
charset-normalizer>=2.1.1
click>=8.1.3
configparser>=5.3.0
coreapi-cli>=1.0.9
coreapi>=2.3.3
coreschema>=0.0.4
cryptography>=37.0.4
cssselect2>=0.6.0
defusedxml>=0.7.1
diff-match-patch>=20200713
dill>=0.3.5.1
django-appconf>=1.0.5
django-background-tasks>=1.2.5
django-cryptography>=1.1
django-easy-pdf>=0.1.1
django-import-export>=2.8.0
django-model-utils>=4.2.0
django-notifications-hq>=1.7.0
django-s3-storage>=0.13.8
django-settings-export>=1.2.1
django-sitetree>=1.17.2
djangorestframework-simplejwt>=5.2.0
dnspython>=2.2.1
enum34>=1.1.10
et-xmlfile>=1.1.0
future>=0.18.2
gunicorn>=20.1.0
h11>=0.13.0
html5lib>=1.1
httplib2>=0.20.4
idna>=3.3
ipaddress>=1.0.23
isort>=5.10.1
itypes>=1.2.0
Jinja2>=3.1.2
jira>=3.4.0
jmespath>=1.0.1
jsonfield>=3.1.0
lazy-object-proxy>=1.7.1
lxml>=4.9.1
MarkupPy>=1.14
MarkupSafe>=2.1.1
mccabe>=0.7.0
oauthlib>=3.2.0
odfpy>=1.4.1
openpyxl>=3.0.10
ordereddict>=1.1
oscrypto>=1.3.0
outcome>=1.2.0
packaging>=21.3
pbr>=5.10.0
pdfrw>=0.4
Pillow>=9.2.0
platformdirs>=2.5.2
psycopg2-binary>=2.9.3
pycparser>=2.21
pyhanko-certvalidator>=0.19.5
pyHanko>=0.13.2
PyJWT>=2.4.0
pylint>=2.14.5
pyparsing>=3.0.9
PyPDF2>=2.10.2
PyPDF3>=1.0.6
pyphen>=0.12.0
PySocks>=1.7.1
python-bidi>=0.4.2
python-dateutil>=2.8.2
python-nmap>=0.6.4
python-taint>=0.42
python-whois>=0.8.0
pytz-deprecation-shim>=0.1.0.post0
pytz>=2022.2.1
PyYAML>=6.0
qrcode>=7.3.1
reportlab>=3.6.11
requests>=2.31.0
requests-oauthlib>=1.3.1
requests-toolbelt>=0.9.1
s3transfer>=0.6.0
selenium>=4.4.3
singledispatch>=3.7.0
six>=1.16.0
sniffio>=1.2.0
sortedcontainers>=2.4.0
soupsieve>=2.3.2.post1
sqlparse>=0.4.2
svglib>=1.4.1
swapper>=1.3.0
tablib>=3.2.1
tinycss2>=1.1.1
toml>=0.10.2
tomli>=2.0.1
tomlkit>=0.11.4
tqdm>=4.64.0
trio-websocket>=0.9.2
trio>=0.21.0
typing_extensions>=4.3.0
tzdata>=2022.2
tzlocal>=4.2
uritemplate>=4.1.1
uritools>=4.0.0
urllib3>=1.26,<2.0
